+++ CHANGELOG.md
@@ -1,7 +1,15 @@
+## 0.3.0
+
+- fixed WebView.storyboard to deployment target 8.0
+- added `InAppBrowser.onLoadResource()` method. The event fires when the InAppBrowser webview loads a resource
+- added `InAppBrowser.addJavaScriptHandler()` and `InAppBrowser.removeJavaScriptHandler()` methods to add/remove javascript message handlers
+- removed `keyboardDisplayRequiresUserAction` from iOS available options
+- now the `url` parameter of `InAppBrowser.open()` is optional. The default value is `about:blank`
+
 ## 0.2.1
 
-- added InAppBrowser.onConsoleMessage() method to manage console messages
-- fixed InAppBrowser.injectScriptCode() method when there is not a return value
+- added `InAppBrowser.onConsoleMessage()` method to manage console messages
+- fixed `InAppBrowser.injectScriptCode()` method when there is not a return value
 
 ## 0.2.0